Added basic framework for rendering world, can make stripy.

+++ ProjectDungeon/src/main/java/com/rollingduck/projectdungeon/world/Coordinates.java
@@ -9,28 +9,28 @@ package com.rollingduck.projectdungeon.world;
  */
 public class Coordinates {
 
-	private int x;
+	private float x;
 
-	private int y;
+	private float y;
 
-	public Coordinates(int x, int y) {
+	public Coordinates(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
